requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
ecdsa==0.18.0
fastapi==0.93.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
